Add new method to UsersProvider and refactor JsonProvider with its interface
